unit of work and repository pattern added

+++ src/infrastructure/data/product_description_entity.py
@@ -2,6 +2,7 @@ from sqlalchemy import Date
 from sqlalchemy import Table
 from sqlalchemy import Column
 from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 from sqlalchemy import String
 from sqlalchemy import ForeignKey
 
@@ -14,7 +15,7 @@ product_description_table = Table(
     Column('name', String(100)),
     Column('category', String(100)),
     Column('subcategory', String(100)),
-    Column('ean', Integer),
+    Column('ean', BigInteger),
     Column('internal_code', String(100)),
     Column('product_dimensions_id', ForeignKey('product_dimensions.id')),
     Column('product_periodicity_id', ForeignKey('product_periodicity.id'))
